config_generator: add consistency check
This commit adds two consistency checks to config_generator tool.
First, we check whether RIB and FIB as reported by show routes
and show forwarding respectively are identical.
Second, we check whether RIFT FIB and kernel FIBs are identical.

Screen scraping for multiple next-hops would be very difficult with
the tables, so I added an CLI command "set cli-mode csv" to specify
that tables should be displayed as CSVs (which is much easier to
parse). In order to a table to be displayed in CSV, one should use
print_table() instead of print().

To make the parsing even easier, I added family options to show
routes and show forwarding, so it's possible to display routes/
forwarding entries for a selected family only.